Final code review
