+++ node-graph/gcore/src/lib.rs
@@ -15,6 +15,7 @@ pub use ctor;
 
 pub mod consts;
 pub mod generic;
+pub mod instances;
 pub mod logic;
 pub mod ops;
 pub mod structural;
